requirements.txt
albumentations==0.5.2
braceexpand==0.1.7
easydict==1.9
imageio==2.34.2
imageio_ffmpeg==0.5.1
imgaug==0.4.0
joblib==1.4.2
kornia==0.5.0
matplotlib==3.9.1
memory_profiler==0.61.0
numpy==2.0.1
omegaconf==2.1.2
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
paddleocr==2.8.1
paddleocr.egg==info
pandas==2.2.2
Pillow==10.4.0
pytorch_lightning==1.2.9
PyYAML==6.0.1
scikit_learn==1.5.1
scipy==1.14.0
six==1.16.0
skimage==0.0
sync_batchnorm==0.0.1
tensorflow==2.17.0
tensorflow_intel==2.17.0
tqdm==4.66.4
webdataset==0.2.86
